AoC Day 2: ugh, what a nightmare
Part one, 16:52; part two, 39:25.

I spent a lot of time trying to narrow the number of options I
needed to consider, but this runs in under a second.  Someday I'll
learn not to prematurely optimize, but Today Is Not That Day.

Switching to a while-loop in part two instead of trying to figure
out how long the "prefix" could be definitely helped too.