Day 4: Well, that was easy...
...mostly because of grid_util.  As they say in Pippin, "It's smarter
to be prepared than it's prepared to be smart."  (They do not.)

Part one, 4:55; part two, 8:26.  With a two-minute-late start because
I was entranced by the Late Show with Stephen Colbert.  The slowdowns
were honestly boneheaded: forgetting to add diagonals=True; forgetting
to check that the square contained an "@" in the first place.

I could imagine optimizing by storing at setup which squares have an @
in them, and iterating only over those, but, well, maybe I did learn
something a couple of days ago about prematurely optimizing?