Set up parse_data() as a no-op
That is, have it return its argument rather than returning nothing.
The point of which is that then parse_data() can be put into both
part_one() and part_two(), and if (as on Day 3) there is no real
data parsing that needs to happen, the no-op will be fine.  But if
there is, it saves having to type "data = parse_data(data)" every
time.

Which is small, but that's the whole point of a template.