AoC Day 6: Hellooooo, zip (and reduce)!
Part one, 9:56; part two, 13:43.

Do enough word puzzles and you can get very, very accustomed to transposing
your list-of-lists matrix, so this was mostly familiar ground.  ("reduce"
less so; I usually just for-loop things, but for once it felt like the right
time to use it.)

Little slowdowns here and there, e.g. forgetting the enumerate() (which is
frustratingly common for me), forgetting to remove the column of spaces in
the second part (leaving them in seemed harmless enough in the first part!).

Of course, Part Two is the entirely of Part One copied over with literally a
single line of code added.  Prospective employers: believe me, I'm better
than this.  (Maybe I'll actually refactor it, for formality's sake.)

(I could also remove, at this point, some of the catches I put in for the
data being weirder than I expected or misparsed, e.g. the "else" block on
determining the operator, or the "[x for x in column if x]" which I put in
to handle a column that had blank space in one or more rows because the
columns weren't all the same length.  But then, I could also not.)