Day 8: Speed up part one, apparently
Really, really would have thought there's a better way than sorting a list
with 999,000 items.  (Probably there is, and it's not what I wrote.)

Also making a long-overdue fix to the template, to print out time taken and
not clock time.  The latter was useful when the program would take far too
long, but it's nice to have sub-second speeds when I've done it right, too.