Day 9: geometry. Why did it have to be geometry?
Part one, 4:50.  Part two...eventually.  (Before midnight.)

I learned my lesson and didn't try to pre-optimize.  As a result, part two
was taking O(heat death of the universe).  Also I didn't read carefully
enough and missed that the interior was also green, so that didn't help, but
honestly, mostly, it was just me having a Very Hard Time conceptualizing.

Finally I thought: what if instead of checking every tile, we just check
every *row* to see whether the entire *range* is green?  We know how to do
ranges!  (That also was too slow, which was when I realized "oh, or we could
check just the four sides, as long as I set up two range-dictionaries".)

So yeah, the point is, I got there.  Eventually.