Day 11: I think my mind's eye has cataracts?
Part 1, 6 minutes; part 2, 21 hours (but in my defense, I spent the
day traveling and visiting with inlaws).

I knew there was just a thing I wasn't quite twigging to.  It was
saving the current path position and the number of things that led
there, rather than the entire path.  If I'd done that, my first
attempt (which is why there are leftover DAC and FFT constants)
might have done fine.  (I may go check that.)